--- app/src/main/java/com/codesoom/assignment/application/UserService.java
@@ -1,117 +1,0 @@
-package com.codesoom.assignment.application;
-
-import com.codesoom.assignment.application.userInterface.UserDeleteInterface;
-import com.codesoom.assignment.application.userInterface.UserFindInterface;
-import com.codesoom.assignment.application.userInterface.UserRegisterInterface;
-import com.codesoom.assignment.application.userInterface.UserUpdateInterface;
-import com.codesoom.assignment.domain.Role;
-import com.codesoom.assignment.domain.RoleRepository;
-import com.codesoom.assignment.domain.User;
-import com.codesoom.assignment.domain.UserRepository;
-import com.codesoom.assignment.dto.UserModificationData;
-import com.codesoom.assignment.dto.UserRegistrationData;
-import com.codesoom.assignment.error.UserEmailDuplicationException;
-import com.codesoom.assignment.error.UserNotFoundException;
-import com.github.dozermapper.core.Mapper;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import javax.transaction.Transactional;
-
-@Service
-@Transactional
-public class UserService implements UserDeleteInterface,
-        UserFindInterface,
-        UserUpdateInterface,
-        UserRegisterInterface {
-    private final Mapper mapper;
-    private final UserRepository userRepository;
-    private final RoleRepository roleRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    public UserService(Mapper dozerMapper,
-                       UserRepository userRepository,
-                       RoleRepository roleRepository,
-                       PasswordEncoder passwordEncoder) {
-        this.mapper = dozerMapper;
-        this.userRepository = userRepository;
-        this.roleRepository = roleRepository;
-        this.passwordEncoder = passwordEncoder;
-    }
-
-    /**
-     * 사용자를 등록한다.
-     *
-     * @param registrationData 등록할 유저의 정보
-     * @throws UserEmailDuplicationException 사용자의 Email이 중복되었을 경우
-     * @return 저장된 유저의 정보
-     */
-    @Override
-    public User registerUser(UserRegistrationData registrationData) {
-        String email = registrationData.getEmail();
-        if (userRepository.existsByEmail(email)) {
-            throw new UserEmailDuplicationException(email);
-        }
-
-        User user = userRepository.save(
-                mapper.map(registrationData, User.class));
-
-        user.changePassword(registrationData.getPassword(), passwordEncoder);
-
-        roleRepository.save(new Role(user.getId(), "USER"));
-
-        return user;
-    }
-
-    /**
-     * 사용자를 수정한다.
-     *
-     * @param id 수정할 사용자의 식별자
-     * @param modificationData 수정할 사용자의 정보
-     * @throws AccessDeniedException 사용자 자원의 식별자와 수정자의 식별자가 다를 경우
-     * @throws UserNotFoundException 식별자에 해당하는 사용자가 존재하지 않을 경우
-     * @return 수정된 사용자
-     */
-    @Override
-    public User updateUser(Long id, UserModificationData modificationData,
-                           Long userId) throws AccessDeniedException {
-        if (!id.equals(userId)) {
-            throw new AccessDeniedException("Access denied");
-        }
-
-        User user = findUser(id);
-
-        User source = mapper.map(modificationData, User.class);
-        user.changeWith(source);
-
-        return user;
-    }
-
-    /**
-     * 사용자를 삭제한다.
-     *
-     * @param id 삭제할 사용자의 식별자
-     * @throws UserNotFoundException 식별자에 해당하는 사용자가 존재하지 않을 경우
-     * @return 삭제된 사용자
-     */
-    @Override
-    public User deleteUser(Long id) {
-        User user = findUser(id);
-        user.destroy();
-        return user;
-    }
-
-    /**
-     * 사용자를 조회한다.
-     *
-     * @param id 조회할 사용자의 식별자
-     * @throws UserNotFoundException 식별자에 해당하는 사용자가 존재하지 않을 경우
-     * @return 식별자에 해당하는 사용자
-     */
-    @Override
-    public User findUser(Long id) {
-        return userRepository.findByIdAndDeletedIsFalse(id)
-                .orElseThrow(() -> new UserNotFoundException(id));
-    }
-}
