+++ app/src/main/java/com/codesoom/assignment/application/user/UserRegisterService.java
@@ -1,0 +1,51 @@
+package com.codesoom.assignment.application.user;
+
+import com.codesoom.assignment.domain.Role;
+import com.codesoom.assignment.domain.RoleRepository;
+import com.codesoom.assignment.domain.User;
+import com.codesoom.assignment.domain.UserRepository;
+import com.codesoom.assignment.dto.UserRegistrationData;
+import com.codesoom.assignment.error.UserEmailDuplicationException;
+import com.github.dozermapper.core.Mapper;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserRegisterService implements UserRegisterInterface {
+
+    private final UserRepository userRepository;
+    private final RoleRepository roleRepository;
+    private final Mapper mapper;
+    private final PasswordEncoder passwordEncoder;
+
+    public UserRegisterService(UserRepository userRepository, RoleRepository roleRepository, Mapper mapper, PasswordEncoder passwordEncoder) {
+        this.userRepository = userRepository;
+        this.roleRepository = roleRepository;
+        this.mapper = mapper;
+        this.passwordEncoder = passwordEncoder;
+    }
+
+    /**
+     * 사용자를 등록한다.
+     *
+     * @param registrationData 등록할 유저의 정보
+     * @throws UserEmailDuplicationException 사용자의 Email이 중복되었을 경우
+     * @return 저장된 유저의 정보
+     */
+    @Override
+    public User registerUser(UserRegistrationData registrationData) {
+        String email = registrationData.getEmail();
+        if (userRepository.existsByEmail(email)) {
+            throw new UserEmailDuplicationException(email);
+        }
+
+        User user = userRepository.save(
+                mapper.map(registrationData, User.class));
+
+        user.changePassword(registrationData.getPassword(), passwordEncoder);
+
+        roleRepository.save(new Role(user.getId(), "USER"));
+
+        return user;
+    }
+}
